+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/LocalDB.Unix.cs
@@ -1,10 +1,10 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
 using System;
 
-namespace Microsoft.Data.SqlClient.SNI
+namespace Microsoft.Data.SqlClient
 {
     internal class LocalDB
     {
