+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/LocalDB.uap.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Data.SqlClient.SNI
+namespace Microsoft.Data.SqlClient
 {
     internal class LocalDB
     {
